feat: Tach /login va /signup servlet, chinh sua lai EmailUtil, ap dung abstraction: trien khai cac interface va cac implement rient biet, co the de dang mo rong code, test nhanh ma khong quan trong logic ben duoi.

+++ src/main/java/nlu/fit/web/souvenirecommerce/controller/admin/AdminProductController.java
@@ -112,7 +112,7 @@ public class AdminProductController extends HttpServlet {
 
             } else if ("edit".equals(action)) {
                 Product product = new Product();
-                product.setId(Integer.parseInt(req.getParameter("id")));
+                product.setId(Long.parseLong(req.getParameter("id")));
                 Category category = new Category();
                 category.setId(Long.parseLong(req.getParameter("categoryId")));
                 product.setCategory(category);
@@ -142,7 +142,7 @@ public class AdminProductController extends HttpServlet {
                 }
 
             } else if ("delete".equals(action)) {
-                int id = Integer.parseInt(req.getParameter("id"));
+                Long id = Long.parseLong(req.getParameter("id"));
 
                 if (productDAO.deleteProduct(id)) {
                     req.setAttribute("message", "Xóa sản phẩm thành công!");
